+++ common_types/src/record_batch.rs
@@ -15,7 +15,7 @@ use arrow_ext::operation;
 use snafu::{ensure, Backtrace, OptionExt, ResultExt, Snafu};
 
 use crate::{
-    column::{cast_nanosecond_to_mills, ColumnBlock, ColumnBlockBuilder},
+    column_block::{cast_nanosecond_to_mills, ColumnBlock, ColumnBlockBuilder},
     datum::DatumKind,
     projected_schema::{ProjectedSchema, RowProjector},
     row::{
@@ -31,7 +31,7 @@ pub enum Error {
     SchemaLen { backtrace: Backtrace },
 
     #[snafu(display("Failed to create column block, err:{}", source))]
-    CreateColumnBlock { source: crate::column::Error },
+    CreateColumnBlock { source: crate::column_block::Error },
 
     #[snafu(display(
         "Failed to create arrow record batch, err:{}.\nBacktrace:\n{}",
@@ -47,7 +47,7 @@ pub enum Error {
     IterateDatum { source: crate::row::Error },
 
     #[snafu(display("Failed to append datum, err:{}", source))]
-    AppendDatum { source: crate::column::Error },
+    AppendDatum { source: crate::column_block::Error },
 
     #[snafu(display(
         "Column not in schema with key, column_name:{}.\nBacktrace:\n{}",
