+++ df_operator/src/udfs/time_bucket.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 CeresDB Project Authors. Licensed under Apache-2.0.
+// Copyright 2022-2023 CeresDB Project Authors. Licensed under Apache-2.0.
 
 //! time_bucket UDF.
 
@@ -6,7 +6,7 @@ use std::time::Duration;
 
 use chrono::{Datelike, FixedOffset, TimeZone};
 use common_types::{
-    column::{ColumnBlock, ColumnBlockBuilder, TimestampColumn},
+    column_block::{ColumnBlock, ColumnBlockBuilder, TimestampColumn},
     datum::{Datum, DatumKind},
     time::Timestamp,
 };
@@ -56,7 +56,9 @@ pub enum Error {
     TruncateTimestamp { timestamp: i64, period: Period },
 
     #[snafu(display("Failed to build result column, err:{}", source))]
-    BuildColumn { source: common_types::column::Error },
+    BuildColumn {
+        source: common_types::column_block::Error,
+    },
 }
 
 define_result!(Error);
