+++ proxy/src/grpc/prom_query.rs
@@ -334,7 +334,7 @@ impl RecordConverter {
 mod tests {
 
     use common_types::{
-        column::{ColumnBlock, ColumnBlockBuilder},
+        column_block::{ColumnBlock, ColumnBlockBuilder},
         column_schema,
         datum::{Datum, DatumKind},
         row::Row,
