+++ table_engine/src/memory.rs
@@ -12,7 +12,7 @@ use std::{
 
 use async_trait::async_trait;
 use common_types::{
-    column::{ColumnBlock, ColumnBlockBuilder},
+    column_block::{ColumnBlock, ColumnBlockBuilder},
     datum::{Datum, DatumKind},
     record_batch::RecordBatch,
     row::{Row, RowGroup},
